Work around quirks in MSVC associated with [d046d5f4c6855].
FossilOrigin-Name: 18b6a43eb8c242e923a852c046a48fe8430d1596c286e114eb1383d6de9adafa

+++ ext/rtree/rtree.c
@@ -634,7 +634,7 @@ static int writeInt64(u8 *p, i64 i){
 static int rtreeValueInt32(sqlite3_value *pVal){
   i64 v64 = sqlite3_value_int64(pVal);
   if( v64>2147483647 ) return 2147483647;
-  if( v64<-2147483648 ) return -2147483648;
+  if( v64<-2147483648LL ) return (int)-2147483648LL;
   return (int)v64;
 }
